+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Changed
 - Save Ensembl ids for demo patient data genes automatically
 - Mock the Ensembl REST API converting gene symbols to Ensembl IDs in tests
+- Changed the tests for the gene conversion Ensembl APIs to use mocks APIs
+- Do not download resourced when testing the update resources command
+- Use requests lib instead of urllib.request to interrogate the Ensembl APIs
 ### Fixed
 - Syntax in the disclaimer
 - Validate the Ensembl gene ID against the Ensembl API when multiple IDs are returned by converting a HGNC symbol.